+++ ecrire/index.php
@@ -139,10 +139,6 @@ elseif (!preg_match(',^[a-z4_][0-9a-z_-]*$,i', $exec)) {
 	set_request('exec', $exec);
 }
 
-// compatibilite ascendante : obsolete, ne plus utiliser
-$GLOBALS['spip_display'] = intval($GLOBALS['visiteur_session']['prefs']['display'] ?? 0);
-$GLOBALS['spip_ecran'] = isset($_COOKIE['spip_ecran']) ? preg_replace('/[^a-z0-9]/i', '', $_COOKIE['spip_ecran']) : 'etroit';
-
 //  si la langue est specifiee par cookie et ne correspond pas
 // (elle a ete changee dans une autre session, et on retombe sur un vieux cookie)
 // on appelle directement la fonction, car un appel d'action peut conduire a une boucle infinie