+++ CHANGELOG.md
@@ -28,4 +28,5 @@
 
 ### Removed
 
+- #5278 Globales obsolètes `spip_ecran` et `spip_display`
 - spip-team/securite#3724 #5150 Suppression de la fonction `initialiser_sel()` (qui ne servait que pour la gestion de htpasswd déportée en plugin).